+++ router_handle/cart.js
@@ -16,11 +16,14 @@ exports.cartHandle = async (req, res) => {
         if (isNaN(quantity)) return res.cc('NaN,该数据类型需为Number')
 
         // 当用户点击添加购物车时，需在数据库中记录该用户的添加的商品
-        // 1.写入购物车表(cart) dql语句
+        // 1.写入购物车表(cart) dql语句 （跟carts里的status状态关联）
         const addDql = `
-            insert into carts (user_id, goods_id, specs, quantity)
-            values (?, ?, ?,?)
-            on DUPLICATE KEY UPDATE quantity = quantity + ?;
+            INSERT INTO carts (user_id, goods_id, specs, quantity)
+            VALUES (?, ?, ?, ?)
+            ON DUPLICATE KEY UPDATE 
+            quantity = CASE WHEN status = 1 THEN VALUES(quantity) ELSE quantity + VALUES(quantity) END,
+            status = 0,
+            deleted_at = NULL
         `
         //     const addDql2 = `
         //     insert into cart_specs (cart_id, spec_value_id,spec_quantity) 
@@ -33,7 +36,7 @@ exports.cartHandle = async (req, res) => {
             sum(c.quantity) as 'cart_total'
             from carts c
             join users u on c.user_id = u.id
-            where u.id = ?
+            where u.id = ? and c.status = 0
         `
         await db.query(addDql, [userId, goodsId, specValueIds, quantity, quantity], (err, results) => {
             if (err) return res.cc(err)
@@ -53,7 +56,7 @@ exports.cartHandle = async (req, res) => {
                 }
                 res.send({
                     status: 0,
-                    msg: '添加成功',
+                    message: '添加成功',
                     data: handleData
                 })
             })
@@ -77,9 +80,7 @@ exports.cartListHandle = async (req, res) => {
             c.quantity,
             gi.url goods_coverImg,
             color_gs.image_url AS color_image,
-            -- 颜色名称（spec_id=1）
             color_gs.value AS color_name,
-            -- 内存名称（spec_id=2）
             memory_gs.value AS memory_name,
             memory_gs.price * c.quantity AS total_price,
               (SELECT price FROM goods_specs
@@ -91,19 +92,19 @@ exports.cartListHandle = async (req, res) => {
             LEFT JOIN goods_specs color_gs
             ON color_gs.id = JSON_UNQUOTE(JSON_EXTRACT(c.specs, '$.color'))
             AND color_gs.goods_id = c.goods_id
-            AND color_gs.spec_id = 1 -- 颜色
+            AND color_gs.spec_id = 1
             LEFT JOIN goods_specs memory_gs
             ON memory_gs.id = JSON_UNQUOTE(JSON_EXTRACT(c.specs, '$.memory'))
             AND memory_gs.goods_id = c.goods_id
-            AND memory_gs.spec_id = 2 -- 内存
-            JOIN goods_images gi on g.id = gi.goods_id
+            AND memory_gs.spec_id = 2
+            JOIN goods_images gi ON g.id = gi.goods_id
             AND gi.type = 'cover'
-            WHERE c.user_id = ?
-            ORDER BY c.id DESC;
+            WHERE c.user_id = ? AND c.status = 0
+            ORDER BY c.id DESC
       `
 
         // 执行查询
-        db.query(sql, [userId], (err, results) => {
+        await db.query(sql, [userId], (err, results) => {
             if (err) return res.cc(err)
 
             res.send({
@@ -111,8 +112,63 @@ exports.cartListHandle = async (req, res) => {
                 data: results
             })
         })
-    } catch (err) {
-        console.error('数据库错误详情:', err)
+    } catch (error) {
+        console.error('数据库错误详情:', error)
         res.send({ status: 1, message: '服务器错误' })
+    }
+}
+
+//更新购物车商品数量
+exports.cartUpdateHandle = async (req, res) => {
+    try {
+        // 获取用户携带过来的购物车id与修改数量
+        const cartId = parseInt(req.body.cartId)
+        const quantity = parseInt(req.body.quantity)
+
+        // 校验
+        if (!cartId || isNaN(cartId)) return res.cc("请传入合法的购物车id")
+        if (!quantity || isNaN(quantity)) return res.cc('数据类型错误，需为Number')
+        // 用户id
+        const userId = req.auth.id
+
+        const dql = 'update carts set quantity = ? where id = ? and user_id = ?'
+        await db.query(dql, [quantity, cartId, userId], err => {
+            if (err) return res.cc('err')
+            res.send({
+                status: 0,
+                message: "succeed"
+            })
+        })
+    } catch (error) {
+        console.error('数据库错误详情:', error)
+        res.cc('服务器错误')
+    }
+}
+
+// 删除购物车对应商品
+exports.cartClearHandle = async (req, res) => {
+    try {
+        // 获取需要删除cartid(是一个数组)
+        const cartIds = req.body.cartIds
+        const userId = req.auth.id
+
+        if(!Array.isArray(cartIds)) return res.cc('数据类型需为数组Array')
+
+        // 占位符数量
+        const placeholders  = cartIds.map(() => '?').join(',')
+        // 这里不使用物理删除（直接删除） 而是逻辑删除（通过status标记）
+        const dql = `update carts set status = 1,deleted_at = NOW()
+                    where user_id = ? and id in (${placeholders}) `
+
+        await db.query(dql, [userId,...cartIds], (err) => {
+            if (err) return res.cc(err)
+            res.send({
+                status: 0,
+                message: 'succeed'
+            })
+        })
+    } catch (error) {
+        console.error('数据库错误详情', error)
+        res.cc('服务器错误')
     }
 }